lapack: add stricter size checks in Dlantr
For certain input (documented as invalid but not checked) DLANTR in
Fortran LAPACK causes out-of-bound writes. Even though this has been
fixed in the Reference and will soon be in OpenBLAS, other
implementations may take longer. Therefore this commit adds checks for
the "invalid" matrix sizes because it's better to panic with a message
rather than cause memory corruption.